Add: ユーザーページのコンポーネントを新規作成し、ヒーローセクションと注目ライブラリを表示
